Homework 8 logs completed

+++ laravel/homeworks/homework8/database/migrations/2024_12_06_124921_create_logs_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('logs', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->dateTime('time');
-            $table->integer('duration');
+            $table->float('duration');
             $table->string('ip', 100)->nullable();
             $table->string('url')->nullable();
             $table->string('method', 10)->nullable();
